feat: Update pull request template
This commit fixes the spacing issue of the pull request checklist by
changing the final checklist item. The relevant information now exists
as a sentence after the list, as well as an additional line.

The additional line, alongside a similar one added to the checklist,
direct the reader to our style guide to review specific guidance that
would otherwise be too verbose to add to the template.